Add structuredstreaming/StructuredUtils.scala
