+++ waterdrop-core/src/main/scala/io/github/interestinglab/waterdrop/output/structuredstreaming/Stdout.scala
@@ -16,20 +16,9 @@ class Stdout extends BaseStructuredStreamingOutputIntra {
   override def getConfig(): Config = config
 
   override def checkConfig(): (Boolean, String) = {
-    if (config.hasPath("triggerMode")) {
-      val triggerMode = config.getString("triggerMode")
-      triggerMode match {
-        case "ProcessingTime" | "Continuous" => {
-          if (config.hasPath("interval")) {
-            (true, "")
-          } else {
-            (false, "please specify [interval] when [triggerMode] is ProcessingTime or Continuous")
-          }
-        }
-        case _ => (true, "")
-      }
-    } else {
-      (true, "")
+    StructuredUtils.checkTriggerMode(config) match {
+      case true => (true, "")
+      case false => (false, "please specify [interval] when [triggerMode] is ProcessingTime or Continuous")
     }
   }
 
@@ -52,21 +41,13 @@ class Stdout extends BaseStructuredStreamingOutputIntra {
       .outputMode(config.getString("outputMode"))
       .option("checkpointLocation", "path/to/HDFS/dir")
 
-    setCheckpointLocation(writer)
+    writer = StructuredUtils.setCheckpointLocation(writer, config)
 
     triggerMode match {
       case "default" => writer
       case "ProcessingTime" => writer.trigger(Trigger.ProcessingTime(config.getString("interval")))
       case "OneTime" => writer.trigger(Trigger.Once())
       case "Continuous" => writer.trigger(Trigger.Continuous(config.getString("interval")))
-    }
-  }
-
-  private def setCheckpointLocation(dw: DataStreamWriter[Row]): DataStreamWriter[Row] = {
-    if (config.hasPath("checkpointLocation")) {
-      dw.option("checkpointLocation", config.getString("checkpointLocation"))
-    } else {
-      dw
     }
   }
 }
